% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.SMARTp}
\alias{print.SMARTp}
\title{Print for SMARTp class}
\usage{
\method{print}{SMARTp}(x, ...)
}
\arguments{
\item{x}{SMARTp object to print}

\item{...}{Other parameters for print}
}
\description{
Print for SMARTp class
}
