% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SMARTp}
\alias{summary.SMARTp}
\title{Summary for SMARTp class}
\usage{
\method{summary}{SMARTp}(object, ...)
}
\arguments{
\item{object}{SMARTp object to summarise}

\item{...}{Other parameters for summary}
}
\description{
Summary for SMARTp class
}
